# crates/client/openapi.yaml
openapi: 3.1.0
info:
  title: Silicon Remind API
  version: 0.1.0
  description: Durable Silicon reminders with organization-wide read access, owner-Silicon mutations,
    IAM application sessions and isolated test environments. Test requests use X-Remind-Test-Key plus
    a test IAM identity; they never fall through to production.
servers:
- url: https://backend.remind.teamofsilicons.com/api/v1
security:
- bearerAuth: []
...
paths:
...
  /test-environments:
    get:
      operationId: listTestEnvironments
      summary: List organization-owned environments
      parameters:
      - *id003
      - *id005
      - *id006
      - name: include_deleted
        in: query
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TestEnvironment'
                  next_cursor:
                    type:
                    - string
                    - 'null'
                required:
                - items
        default: *id001
    post:
      operationId: createTestEnvironment
      summary: Create an empty IAM-bound test environment
      parameters:
      - *id003
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentCreated'
        default: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTestEnvironment'
...
  /testing-environment:
    get:
      operationId: currentEnvironment
      summary: Describe the selected sandbox
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestEnvironment'
        default: *id001
      security:
      - testKeyAuth: []
      description: Requires the Remind environment key only. Omitting it returns test_environment_required.
        Cleaning does not clean the linked IAM environment. A sandbox permits at most 100 retained reminders;
        inactivity retires it after 15 days.
  /testing-environment/cleanings:
    post:
      operationId: cleanEnvironment
      summary: Atomically clear all sandbox data, keeping the IAM binding and root key
      parameters: []
      responses:
        '204':
          description: Success
        default: *id001
      security:
      - testKeyAuth: []
      description: Requires the Remind environment key only. Omitting it returns test_environment_required.
        Cleaning does not clean the linked IAM environment. A sandbox permits at most 100 retained reminders;
        inactivity retires it after 15 days.
...
  /testing-environment/iam:
    put:
      operationId: configureEnvironmentIam
      summary: Install or rotate the sandbox IAM Application credential
      security: []
      parameters:
      - $ref: '#/components/parameters/TestKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - iam_app_secret
              properties:
                iam_app_secret:
                  type: string
                  writeOnly: true
      responses:
        '204':
          description: Credential verified in the bound IAM test world and encrypted at rest
        default:
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: IAM Remind-audience application token
    testKeyAuth:
      type: apiKey
      in: header
      name: X-Remind-Test-Key
      description: 32-character Remind root key. Selects an isolated test environment; ordinary actions
        still require a test IAM access token.
  parameters:
...
    TestKey:
      name: X-Remind-Test-Key
      in: header
      required: false
      description: Optional sandbox selector. Invalid, expired or revoked keys fail closed. Environment
        management routes reject this header.
      schema:
        type: string
        pattern: ^[a-zA-Z0-9]{32}$
  responses:
...
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - request_id
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
...
    TestEnvironment:
      type: object
      properties:
        id: *id007
        org_id: *id009
        creator_id: *id007
        name: *id009
        description:
          type:
          - string
          - 'null'
        iam_environment_id: *id007
        version:
          type: integer
        created_at: *id011
        last_activity_at: *id011
        deleted_at: &id012
          type:
          - string
          - 'null'
          format: date-time
        purge_after: *id012
      required:
      - id
      - org_id
      - creator_id
      - name
      - description
      - iam_environment_id
      - version
      - created_at
      - last_activity_at
      - deleted_at
      - purge_after
    CreateTestEnvironment:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type:
          - string
          - 'null'
          maxLength: 10000
        iam_test_key:
          type: string
          pattern: ^[a-zA-Z0-9]{32}$
          writeOnly: true
        iam_app_secret: *id002
      required:
      - name
      - iam_test_key
    EnvironmentCreated:
      type: object
      properties:
        environment:
          $ref: '#/components/schemas/TestEnvironment'
        key:
          type: string
          pattern: ^[a-zA-Z0-9]{32}$
      required:
      - environment
      - key
...